move register and login to users folder in components
